requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
asttokens==3.0.0
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bibtexparser==1.4.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
docutils==0.21.2
executing==2.1.0
fake-useragent==2.0.3
fastapi==0.115.6
free_proxy==1.1.3
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imagesize==1.4.1
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonpickle==4.0.1
loguru==0.7.3
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
networkx==3.4.2
numpy==2.2.1
openai==1.58.1
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
PubChemPy==1.0.4
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyMuPDF==1.25.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyvis==0.3.2
rdkit==2024.3.6
redis==5.0.1
requests==2.32.3
scholarly==1.7.11
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.41.3
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
